rename actions name
